refactor(user): reorganize user type definitions and hooks
Move IndexedUser interface to user subdirectory and extract artist request mutation hook
Remove duplicate IndexedUser type definition and update imports
