+++ frontend/src/api/user/auth/mutations.ts
@@ -5,11 +5,6 @@ import { useNavigate } from "react-router-dom";
 import { useMutation } from "@tanstack/react-query";
 import { logout, setUser } from "../../../redux/slices/userSlice";
 
-interface ArtistRequestPayload {
-  phone?: string;
-  bio?: string;
-  country?: string;
-}
 
 
 
@@ -22,19 +17,6 @@ interface ArtistRequestPayload {
 
 
 
-export const useCreateArtistRequestMutation = () => {
-  return useMutation({
-    mutationFn: (data: ArtistRequestPayload) =>
-      apiClient.post("/api/v1/user/artist-request", data),
-    onSuccess: (res) => {
-      console.log("Artist request submitted:", res.data);
-      toast.success("Artist request submitted successfully!");
-    },
-    onError: (error: any) => {
-      toast.error(error?.response?.data?.message || "Something went wrong!");
-    },
-  });
-};
 
 export const useLogoutMutation = () => {
   const dispatch = useDispatch();
