+++ frontend/src/features/user/components/auth/BecomeAnArtist.tsx
@@ -5,7 +5,7 @@ import { Input } from "../../../../components/ui/input";
 import { Textarea } from "../../../../components/ui/textarea";
 import type { RootState } from "../../../../redux/store";
 import type { User } from "../../../../types/user/user";
-import { useCreateArtistRequestMutation } from "../../../../api/user/auth/mutations";
+import { useCreateArtistRequestMutation } from "../../hooks/art/useCreateArtistRequestMutation";
 import toast from "react-hot-toast";
 import { z } from "zod";
 import { useHasSubmittedArtistRequest } from "../../../../api/user/art/queries";
