+++ frontend/src/types/art/artistRequestPayload.ts
@@ -1,0 +1,5 @@
+export interface ArtistRequestPayload {
+  phone?: string;
+  bio?: string;
+  country?: string;
+}
